\name{bee2r-package}
\alias{bee2r-package}
\alias{bee2r}
\docType{package}
\title{
\packageTitle{bee2r}
}
\description{
\packageDescription{bee2r}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bee2r}
\packageIndices{bee2r}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bee2r}

Maintainer: \packageMaintainer{bee2r}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
